temp

+++ modules/core/media/interfaces/imediainfo.h
@@ -20,7 +20,7 @@ protected:
 
     int channels;
     int bitrate;
-    int duration;
+    qint64 duration;
     int sampleRate;
     qint64 size;
 
@@ -54,7 +54,7 @@ public:
     inline int getTrack() const { return track; }
     inline int getChannels() const { return channels; }
     inline int getBitrate() const { return bitrate; }
-    inline int getDuration() const { return duration; }
+    inline qint64 getDuration() const { return duration; }
     inline int getSampleRate() const { return sampleRate; }
 
     inline void setSize(qint64 newSize) { size = newSize; }
